refactor(menubar): extract Notifier/NotifierDelegate into Notifier.swift (PHNX-4004)
Pure move of NotifierDelegate, Notifier (post/runOneShot/clickURL/
wireClickHandler) out of PromptPanel.swift into its own file. No behavior
change — the next commit rewrites it onto UNUserNotificationCenter.
